chore: Update deployment and Docker configurations; enhance response model
